mergeable(): move inside softDeleted()
