+++ apps/_back02keystone/custom-plugins/index.js
@@ -2,7 +2,6 @@ const { uuided } = require('./uuided')
 const { versioned } = require('./versioned')
 const { tracked } = require('./tracked')
 const { softDeleted } = require('./softDeleted')
-const { mergeable } = require('./mergeable')
 const { historical } = require('./historical')
 
 module.exports = {
@@ -10,6 +9,5 @@ module.exports = {
     versioned,
     tracked,
     softDeleted,
-    mergeable,
     historical,
 }
